refactor(rimap-audit): extract writer/core.rs from writer/mod.rs
Moves AuditOptions, AuditWriter, Inner, FailureInjection, the
impl AuditWriter block (open/path/rotate_bytes/process_id/
suppressed_failures/force_next_write_failure/bytes_written/
on_disk_len), set_file_mode_0600, set_parent_mode_0700, and the
tests module from writer/mod.rs (1003 LOC) to a new writer/core.rs
sibling (989 LOC). writer/mod.rs shrinks to 23 LOC of pure wiring +
re-exports.

AuditWriter's fields are bumped from private (file-scope) to
pub(super) so siblings (emit.rs, log.rs, rotation.rs) that already
mutate them via `self.path` / `self.inner.lock_inner` etc. keep
working without going through accessor methods. Same widening for
FailureInjection.fail_next. No public-API change; pub(crate) and
pub re-exports from mod.rs preserve every external import path.

No behavior change. All 1416 workspace tests still pass.

Refs #300 (item 3, hub 1/4)
